added dtos for operations, organized controllers, added functionalities

+++ src/main/java/com/example/yonetimSistemi/dataAccess/abstracts/DepartmanRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.yonetimSistemi.entities.Departman;
 
-public interface DepartmanRepository extends JpaRepository<Departman, Long>{
+public interface DepartmanRepository extends JpaRepository<Departman, Integer>{
 	
 	Departman findByDepartmanAdi(String departmanAdi);
